[#77] refactor : 팀 생성 폴더 구조 바꾸기
